Added users

+++ WoWRaidScore/tasks.py
@@ -19,9 +19,11 @@ def mul(x, y):
 def xsum(numbers):
     return sum(numbers)
 
+
 @shared_task()
-def parse_task(raid_id):
-    parse_raid_task(raid_id)
+def parse_task(raid_id, user, group):
+    parse_raid_task(raid_id, user, group)
+
 
 def update_status(progress, update_progress=True):
     if update_progress:
@@ -32,18 +34,21 @@ def get_progress(current_analyzer, num_analyzers):
     return int(float(current_analyzer) / num_analyzers * 90.0)
 
 
-def parse_raid_task(raid_id, update_progress=True):
+def parse_raid_task(raid_id, user, group, overwrite=True, update_progress=True):
     wcl_client = WCLRequests(raid_id)
     update_status(0.0, update_progress)
     try:
         raid = Raid.objects.get(raid_id=raid_id)
-        for fight in Fight.objects.filter(raid=raid):
-            for raid_score in RaidScore.objects.filter(fight=fight):
-                raid_score.delete()
-            fight.delete()
+        if overwrite:
+            for fight in Fight.objects.filter(raid=raid):
+                for raid_score in RaidScore.objects.filter(fight=fight):
+                    raid_score.delete()
+                fight.delete()
+        else:
+            raise Exception("The raid has already been parsed")
     except ObjectDoesNotExist:
         wcl_raid = wcl_client.get_raid_info()
-        raid = Raid(raid_id=raid_id, time=wcl_raid.start_time)
+        raid = Raid(raid_id=raid_id, time=wcl_raid.start_time, user=user, group=group)
         raid.save()
     update_status(2.0, update_progress)
     fights = wcl_client.get_fights()
